requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astroid==3.3.10
astunparse==1.6.3
beautifulsoup4==4.13.4
black==25.1.0
boto3==1.38.18
botocore==1.38.18
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
coverage==7.3.2
cycler==0.12.1
dill==0.4.0
Django==4.2.9
django-debug-toolbar==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
filelock==3.18.0
flake8==7.2.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==21.2.0
h5py==3.13.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
mypy_extensions==1.1.0
namex==0.0.9
networkx==3.4.2
numpy==2.1.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
s3transfer==0.12.0
scipy==1.15.3
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tomlkit==0.13.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.139
ultralytics-thop==2.0.14
urllib3==2.4.0
vercel==0.2.1
Werkzeug==3.1.3
whitenoise==6.5.0
wrapt==1.17.2
